added WebSockets and static pages for testing WebSockets

+++ GolangHackatonAuth/controllers/plant.go
@@ -12,13 +12,33 @@ type PlantController struct {
 	beego.Controller
 }
 
+// @securityDefinitions.apikey BearerAuth
+// @in header
+// @name Authorization
+func (p *PlantController) checkAuth() bool {
+	auth := AuthController{Controller: p.Controller}
+	if auth.CheckAuth() {
+		p.Data["json"] = Respons{Err: true, Data: "Unauthorized"}
+		p.Ctx.Output.SetStatus(401)
+		p.ServeJSON()
+		return false
+	}
+	return true
+}
+
 // @Title CreatePlant
 // @Description create plants
+// @Security BearerAuth
+// @Param Authorization header string true "Insert your access token" default(Bearer <Add access token here>)
 // @Param	body		body 	models.Plant	true		"body for plant content"
 // @Success 200 {int} models.Plant.Id
 // @Failure 403 body is empty
 // @router / [post]
 func (p *PlantController) Post() {
+	if !p.checkAuth() {
+		return
+	}
+
 	var plant models.Plant
 	if err := json.Unmarshal(p.Ctx.Input.RequestBody, &plant); err != nil {
 		p.Data["json"] = Respons{Err: true, Data: "Invalid JSON format"}
@@ -37,10 +57,14 @@ func (p *PlantController) Post() {
 
 // @Title GetAll
 // @Description get all plants
+// @Security BearerAuth
 // @Param Authorization header string true "Insert your access token" default(Bearer <Add access token here>)
 // @Success 200 {object} models.Plant
 // @router / [get]
 func (p *PlantController) GetAll() {
+	if !p.checkAuth() {
+		return
+	}
 	plants := models.GetAllPlants()
 	p.Data["json"] = Respons{Err: false, Data: plants}
 	p.ServeJSON()
@@ -48,11 +72,15 @@ func (p *PlantController) GetAll() {
 
 // @Title Get
 // @Description get plant by pid
+// @Security BearerAuth
 // @Param	pid		path 	string	true		"The key for staticblock"
 // @Success 200 {object} models.Plant
 // @Failure 403 :pid is empty
 // @router /:pid [get]
 func (p *PlantController) Get() {
+	if !p.checkAuth() {
+		return
+	}
 	pid, err := p.GetInt(":pid")
 	if err == nil {
 		plant, err := models.GetPlant(pid)
@@ -67,11 +95,15 @@ func (p *PlantController) Get() {
 
 // @Title Update
 // @Description update the plant
+// @Security BearerAuth
 // @Param	body		body 	models.Plant	true		"body for plant content"
 // @Success 200 {object} models.Plant
 // @Failure 403 body is empty
 // @router / [put]
 func (p *PlantController) Put() {
+	if !p.checkAuth() {
+		return
+	}
 	var plant models.Plant
 	json.Unmarshal(p.Ctx.Input.RequestBody, &plant)
 	err := models.UpdatePlant(&plant)
@@ -85,11 +117,15 @@ func (p *PlantController) Put() {
 
 // @Title Delete
 // @Description delete the plant
+// @Security BearerAuth
 // @Param	pid		path 	string	true		"The pid you want to delete"
 // @Success 200 {string} delete success!
 // @Failure 403 pid is empty
 // @router /:pid [delete]
 func (p *PlantController) Delete() {
+	if !p.checkAuth() {
+		return
+	}
 	pid, err := p.GetInt(":pid")
 	if err == nil {
 		del := models.DeletePlant(pid)
